examples/rag/crag_service/requirements_crag.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langgraph==0.0.20
chromadb==0.4.22
tiktoken==0.5.2
tavily-python==0.3.0
requests==2.31.0
unstructured==0.12.0
markdown==3.5.1
